requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
requests==2.32.4
faiss-cpu==1.11.0.post1
sentence-transformers==5.1.0
tree-sitter==0.25.1
tree-sitter-languages==1.10.2
numpy==2.3.3
